# go/document.go
/*
 * Screenwriter API
 *
 * Data API for Screenwriter: an online collaborative screenwriting software.
 *
 * API version: 0.0.1
 * Contact: logan@theinitiativepg.ocm
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Document struct {

	Id int64 `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	Type_ string `json:"type,omitempty"`

	Text string `json:"text,omitempty"`

	Author int64 `json:"author,omitempty"`

	CreatedDate string `json:"created_date,omitempty"`

	UpdatedDate string `json:"updated_date,omitempty"`
}